Python版への移植
